v2

+++ database/migrations/2025_06_08_233725_create_discounts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('discounts', function (Blueprint $table) {
             $table->id();
-            $table->integer('discount');
+            $table->decimal('discount');
             $table->timestamps();
         });
     }
